Refactor : User클래스 수정
