+++ server/src/main/java/seb39_40/coffeewithme/user/domain/UserStatus.java
@@ -1,0 +1,15 @@
+package seb39_40.coffeewithme.user.domain;
+
+import lombok.Getter;
+
+public enum UserStatus {
+    USER_SIGNUP("signup"),
+    USER_WITHDRAW("withdraw");
+
+    @Getter
+    private String status;
+
+    UserStatus(String status) {
+        this.status = status;
+    }
+}
